generator in back_tracking opti

+++ generateur/generator.h
@@ -20,9 +20,8 @@ typedef	struct	s_map
   int		width;
   int		height;
   int		line;
-  int		index;
+  short int	index;
 }		t_map;
-
 
 int	check_dead_end(t_map *map, int coord);
 t_map	*create_map(int height, int width);
